fix memory leak since CC 5.7.2
Since we upgraded to concourse 5.7.2 it seems the tool create some memory leak
we updated the code to prevent it by using the new CC way to open db conn.

We also made persistent opened connections to the database.

+++ build.sh
@@ -2,6 +2,11 @@
 
 export PATH=$PATH:$GOPATH/bin
 
+if [ -z "$GOPATH" ]; then
+  echo "Ensure GOPATH is set"
+  exit 1
+fi
+
 # Cleanup
 rm -rf $GOPATH/bin
 
@@ -9,4 +14,4 @@ rm -rf $GOPATH/bin
 #CGO_ENABLED=0 GOOS=linux go build -o bin/concourse-toolkit main.go
 
 # build
-go build -o bin/concourse-toolkit main.go
+GOFLAGS=-mod=vendor go build -o bin/concourse-toolkit main.go
